# contracts/openapi/console-api.yaml
openapi: 3.0.3
info:
  title: Dynamiq Console API
  description: >
    Control-plane API served by console-api (Phalcon). PRD-001 §9.1.
    Every /admin/v1/* route except /me and /companies requires the X-Company-Id header.
  version: "0.1.0"
servers:
  - url: /admin/v1
security:
  - ssoJwt: []
components:
  securitySchemes:
    ssoJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: SSO JWT issued by the existing Phalcon system (RS256 + JWKS), PRD §7.1/7.6.
  parameters:
    XCompanyId:
      name: X-Company-Id
      in: header
      required: true
      schema:
        type: string
      description: Company context for this request; checked against membership (PRD §7.2).
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Connection:
      type: object
      properties:
        id: { type: string }
        company_id: { type: string }
        name: { type: string }
        type: { type: string, example: dynamiq.connections.OpenAI }
        api_base: { type: string }
        masked_hint: { type: string }
        max_concurrency: { type: integer }
...
    Tool:
      type: object
      description: >-
        CRUD + standalone connectivity test only (PRD §6.5) — not yet wired into the agent
        spec/compiler, so no agent can call one during a run.
      properties:
        id: { type: string }
        company_id: { type: string }
        name: { type: string }
        kind: { type: string, enum: [http, builtin, python] }
        access_level: { type: string, enum: [read, write] }
        auth_mode: { type: string, enum: [service, delegated] }
        audience: { type: string, nullable: true }
        config:
          type: object
          additionalProperties: true
          description: 'kind-specific: for http, url/method/headers/description/input_schema.'
        enabled: { type: boolean }
...
paths:
...
  /connections/{id}:
    parameters:
      - { $ref: '#/components/parameters/XCompanyId' }
      - { $ref: '#/components/parameters/IdParam' }
    get:
      operationId: getConnection
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Connection' } } } }
    patch:
      operationId: updateConnection
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Connection' }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Connection' } } } }
    delete:
      operationId: deleteConnection
      responses:
        "204": { description: Deleted }
...
  /tools:
    get:
      operationId: listTools
      parameters: [{ $ref: '#/components/parameters/XCompanyId' }]
      responses: { "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Tool' } } } } } }
    post:
      operationId: createTool
      parameters: [{ $ref: '#/components/parameters/XCompanyId' }]
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/Tool' } } } }
      responses: { "201": { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Tool' } } } } }
  /tools/{id}:
    parameters:
      - { $ref: '#/components/parameters/XCompanyId' }
      - { $ref: '#/components/parameters/IdParam' }
    patch:
      operationId: updateTool
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/Tool' } } } }
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Tool' } } } } }
    delete:
      operationId: deleteTool
      responses: { "204": { description: Deleted } }
  /tools/{id}/test:
    parameters:
      - { $ref: '#/components/parameters/XCompanyId' }
      - { $ref: '#/components/parameters/IdParam' }
    post:
      operationId: testTool
      summary: >-
        Proves a `kind: http` tool's configured URL is reachable past the egress allowlist
        (PRD §6.5/§7.4) — not yet an authenticated call, see ai-internal.yaml's testTool.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  detail: { type: string }
...